fix: rename entity file and fix some import issues
