+++ src/plugins/datasource.ts
@@ -12,7 +12,7 @@ import { AppSetting } from '../services/item/plugins/app/appSetting/appSettings'
 import { App } from '../services/item/plugins/app/entities/app';
 import { Publisher } from '../services/item/plugins/app/entities/publisher';
 import { ItemGeolocation } from '../services/item/plugins/geolocation/ItemGeolocation';
-import { Invitation } from '../services/item/plugins/invitation/invitation';
+import { Invitation } from '../services/item/plugins/invitation/entity';
 import { Category } from '../services/item/plugins/itemCategory/entities/Category';
 import { ItemCategory } from '../services/item/plugins/itemCategory/entities/ItemCategory';
 import { ItemFavorite } from '../services/item/plugins/itemFavorite/entities/ItemFavorite';
