+++ src/services/item/plugins/invitation/repository.ts
@@ -3,8 +3,8 @@ import { ResultOf } from '@graasp/sdk';
 import { AppDataSource } from '../../../../plugins/datasource';
 import { Member } from '../../../member/entities/member';
 import { mapById } from '../../../utils';
+import { Invitation } from './entity';
 import { InvitationNotFound } from './errors';
-import { Invitation } from './invitation';
 
 /**
  * Database's first layer of abstraction for Invitations
